mpris: Fix double free
When the dbus name cannot be acquired, name_lost_handler() is called once
during startup and again during shutdown. This causes the unregister
functions, which free the signal id arrays, to be called twice in a row.
This commit make the unregister functions set the pointers to signal id
arrays to NULL when freed and check that the pointers are not NULL
before doing anything else.