Improve mpv error message handling
Previously, log messages from mpv are shown to the user in a modal
dialog box on a per-message basis. This bahavior can be annoying at
times since mpv sometimes produce large amount of repeated error
messages in rapid sucession. The new implementation uses
mpv_event_end_file.error to determine what and if an error message needs
to be presented to the user. Users will only be notified of errors that
are serious enough that playback cannot continue.

The gschema key mpv-msg-redir-enable is also removed. All mpv log
messages are now redirected to stdout.