Simplify pref_dialog construction and preferences handling
Encode the content of pref_dialog in an array and construct it by going
through the array and adding new widgets one-by-one based on the encoded
information. Bindings to GSettings keys are established using
g_settings_bind() for most widget types. File choosers listen to
file-set signal and update the keys manually since there is no suitable
property to bind with.