pref_dialog: Fix missing space when CSD is disabled
When CSD is disabled, pref_dialog needs some extra space over the action
area. Previously, NULL-ness of the header bar was incorrectly used to
determine if CSD is enabled. However, the header bar is never NULL in
the init function. The extra space is now added in
pref_dialog_constructed() using use-header-bar to determine if CSD is
enabled.