main_window: Fix incorrect playlist size
If playlist is hidden in the saved window state, the size of the
playlist will be larger than last saved size when the user makes it
visible. The increased size is caused by incorrect GtkPaned handle
position. Fix this by making main_window_set_playlist_visible() set the
handle position to the right edge of the window before showing the
playlist.