application: Fix crash when opening files without dot in the name
The subtitle detection code extracts the extension from the file name
by using strrchr() to find the last dot in the filename and extract the
substring past that point. When the name doesn't contain a dot, strrchr()
returns NULL. This is then incremented by 1, which normally moves the
pointer past the dot. However, when the returned pointer is NULL, the
resulting pointer points to 0x1 instead, which crashes the program when
it is dereferenced.

This only occurs when opening files using DnD, since subtitle detection
is not done when opening files in other ways.