mpv_obj: Add mpv_obj_load_list()
The function loads a list of uris by internally looping through it and
load each of the uri using mpv_obj_load(). This also fixes the problem
where files appended to the playlist using the context menu will fail to
play automatically if the playlist was cleared manually prior to loading
the file.