application: Handle uninitialized MpvObj in open_handler()
When the GAction 'open' is activated, the MpvObj may not have been
initialized yet. Handle this correctly by adding 'ready' to MpvObjState.
This variable will be true only if MpvObj is initialized. If MpvObj is
uninitailized at the time 'open' is activated, wait until 'mpv-init' is
fired before loading the files.