Bump minimum GTK+ version to 3.20 and remove --disable-opengl-cb
GTK+ 3.20 fixes a bug in window resizing that was previously worked
around by gnome-mpv. Updating the code to work properly with GTK+ 3.20
would mean that gnome-mpv would no longer work properly with older GTK+
versions. Since --disable-opengl-cb was added to allow building with
GTK+ 3.14, it is no longer useful to keep it.