main_window: Add main_window_resize_video_area()
This function resizes the window so that the video area will be of
specified size. This is done by first naively resizing the window to the
given size. This will almost certainly result in the wrong size for the
video area. If this is the case, then the difference between the target
size and the actual size of the video area is used as offsets for
another resize, which should now give correct size for the video area.

Also update resize_window_to_fit() to use this function, and remove
main_window_get_width_margin() and main_window_get_height_margin() as
they are no longer necessary.