main_window: Fix main_window_resize_video_area()
Only adjust resize offset when the target size is smaller than the
screen, and the window is not maximized or fullscreened. This fixes the
problem where resizing the window to a size larger than te screen causes
the resize offset to drift so far from the correct value that more than
one call to main_window_resize_video_area() may be required to get the
correct size.