main_window: Fix playlist size when the window is maximized
When the playlist is shown for the first time while the window is
maximized, the size of the playlist will be incorrect; the handle of
GtkPaned will be placed at the right border of the window prior to
maximization. Fix this by moving the handle to the correct position in
that situation.