mpris: Improve metadata handling
Directly retrieve "metadata" property as MPV_FORMAT_NODE and iterate
through the result rather than trying to retrieve tags from a small set
of known tags. This allows custom tags to be exported to MPRIS clients
and eliminates log messages about failing to retrieve tags that are
unavailable in the current file.

Fixes #164.