application: Reset keys when the main window is shadowed
When a keybinding results in a modal dialog being shown, the key release
event is delivered to the dialog rather than the main window. Those keys
would then be stuck in the down state until the keys are pressed again
after the main window is unshadowed. Fix this by sending a keyup command
with no argument to mpv when the main window is shadowed, which will set
all keys to the up state.