def: Override mpv bindings for seeking and adjusting volume with a mouse
By default, mpv maps scrolling (MOUSE_BTN3/4) to seeking and tilting the
scroll wheel (MOUSE_BTN5/6) to adjusting volume. Reverse this so that
seeking is done by tilting the scroll wheel and adjusting the volume is
done by scrolling instead. This is better because volume can also be
adjusted by hovering on the volume button and scrolling. If the user
misses the button, it can cause the player to seek instead, which can be
annoying especially when streaming.