mpv: Prevent vo detection inconsistency
When files are loaded very early from a lua script, the current-vo
remains NULL right after initialization even if non-opengl-cb vo is
chosen. This causes GmpvMpv to misdetect the vo as opengl-cb during
initialization, but later when the property becomes available, the
detection will be again correct. This commit adds a new variable that
stores the result of detection during initialization and adds a new
function that returns this result, which prevents inconsistency from
arising. However, it doesn't fix the actual misdetection.