Handle files loaded before VO is configured
When a file is loaded before the vo is configured, which may happen when
the loadfile command is run from a Lua script, playback will fail. Fix
this by saving mpv's playlist on MPV_EVENT_START_FILE if the vo is not
yet configured at that point and later load the files in the saved
playlist when the vo becomes ready.

Fixes #268.