mpris: Fix crash when failing to own MPRIS bus name
The GmpvMprisModule passed to gmpv_mpris_module_register_interface() and
gmpv_mpris_module_unregister_interface() may be NULL when failing to own
the bus name. When this happens, calling GMPV_MPRIS_MODULE_GET_CLASS()
on it will lead to a crash. Fix this by only attempting to get the class
of the module only if it is non-NULL.

Fixes #281.