video_area: Correctly handle toggling controls visibility in fullscreen
When the controls is made visible by toggling its visibility (e.g. by
pressing Ctrl+H), the bottom margin of the video area is set to
non-zero because the function did not take into the account the fact
that the controls is always in floating mode if the fullscreen mode is
active. Fix this by ensuring that the bottom margin is always zero in
fullscreen mode.