Create an extra instance of GmpvControlBox for GmpvVideoArea
This makes it possible to avoid reparenting GmpvControlBox when entering
or leaving fullscreen mode, which often causes assertion failure
gtk_container_propagate_draw() when a video is playing when opengl-cb is
used. This is probably due to event queue congestion that causes the
draw event to be processed after the control box is moved.