application: Fix extra window when opening files from CLI
When the always-open-new-window option is set, and the application is
launched from CLI with one or more files, two windows are created. This
is because command_line_handler() will emit an activate signal causing a
window to be created upon starting. After that, open_handler() will
again emit a new-window signal, creating another window. Fix this by
checking the number of files to be opened in command_line_handler() and
only emit the activate signal or the new-window action if there are no
files to be opened or --new-window is set.

Fixes #318.