Update mpv's window-scale property when window is resized
mpv only fires property change event if the actual value of the property
changes. This means that if the window-scale property is set to some
value, the window is resized by the user, and the property is again set
to the same value, no event will be fired. Fix this by keeping the mpv
property consistent with the actual size of the window.