video-area: Add support for dead zone
Add GSettings option for configuring dead zone, an area in which mouse
movement will not cause the controls to be shown. The width of the dead
zone is always equal to the width of the video area. The height is
calculated by multiplying the height of the video area with the value of
the option. The dead zone is centered vertically so that there's equal
amount of "active zone" on either side of it.

Fixes #515.