controller: Fix CelluloidModel not initializing on X11
On X11, CelluloidView could be realized before it is manually shown.
This causes the handler for the ready signal to never fire, which means
that CelluloidModel never gets initialized. Fix this by checking if
CelluloidView is already initialized before showing it, if so, call the
ready signal handler manually.

Fixes #661.