mpris: Initialize interfaces before acquiring name
Some desktop applications look for media players by listening to the
DBus NameOwnerChanged signal which is emitted after a name is acquired
and match on a well-known name prefix. If the MPRIS interfaces are not
initialized at this time, it may lead to subtle bugs. Initialize the
interfaces before the name is acquired to ensure the interface is
never partially implemented.