controller: Fix autofit triggering regardless of settings
When a video is opened for the first time, the window-scale property is
updated, triggering an autofit. Fix this by checking if window-scale is
transitioning from the value of zero, which indicates that a video is
being opened for the first time, and only requesting an autofit if this
is not the case.

Fixes #705.