video-area: Fix cursor autohiding outside of video area
With the migration to GTK4,the entire window shares a single GdkSurface.
Because of this, when the video area hides the cursor, it gets hidden
everywhere, even outside of the video area. Fix this by checking if the
cursor is hovering the video area before trying to hide it.

Fixes #717.