def: Map Ctrl+Q to the quit command
The previous mapping used script message to trigger a GAction to close
the window, but this somehow triggers a crash when using Flatpak. This
was done to workaround an issue where the window size wouldn't be saved
if the quit command was used, but this is no longer an issue.

Fixes #741.