controller: Use gtk_window_destroy() to destroy CelluloidView
Previously, views were being destroyed with g_object_unref(), which
doesn't remove the view from the list of toplevels. Because of this, a
stale pointer could be left in the list of toplevels after a window is
closed. This can lead to a crash if, for example, GTK references the
stale pointer afterwards.

Fixes #799.