file-chooser: Allow disabling state restoration
Add a new parameter to celluloid_file_chooser_new() that controls
whether or not the state, i.e. the last opened folder, would be
restored. This has priority over the last-folder-enable GSettings key,
meaning both the parameter and the GSettings key have to be set to TRUE
for the last folder to be restored.