video-area: Inhibit eager reveal of floating controls
This commit inhibits Celluloid from revealing floating controls upon
launching a video or fullscreen toggle. The floating controls should now
be revealed only upon cursor motion.

The old behaviour was rather annoying, especially when the user opened
or fullscreened a video using the keyboard (pressing Enter in Nautilus,
pressing 'f'); Once the Celluloid window opened or fullscreened in this
manner, not using the mouse (cursor) would not trigger the reveal
timeout, which would keep the floating controls revealed indefinitely,
forcing the user to wiggle the cursor to trigger the timeout to hide the
floating controls.

It was also not often the desired default behaviour to even see the
floating controls upon video launch at all. Rather, one might opine that
a keybinding for revealing the floating controls would be a better UX.
Perhaps something like Shift or Alt.